# feat(source/istio): reduce cpu cycles with indexers