--- a/source/endpoint_benchmark_test.go
+++ b/source/endpoint_benchmark_test.go
@@ -28,6 +28,7 @@ import (
 	kubeinformers "k8s.io/client-go/informers"
 	coreinformers "k8s.io/client-go/informers/core/v1"
 	"k8s.io/client-go/kubernetes/fake"
+	"sigs.k8s.io/external-dns/source/informers"
 
 	v1alpha3 "istio.io/api/networking/v1alpha3"
 	istiov1a "istio.io/client-go/pkg/apis/networking/v1"
@@ -92,14 +93,9 @@ func svcInformerWithServices(toLookup, underTest int) (coreinformers.ServiceInfo
 	svcInformer := informerFactory.Core().V1().Services()
 	ctx := context.Background()
 
-	_, err := svcInformer.Informer().AddEventHandler(
-		cache.ResourceEventHandlerFuncs{
-			AddFunc: func(obj interface{}) {
-			},
-		},
-	)
+	err := svcInformer.Informer().AddIndexers(informers.IndexerSpecSelector[*corev1.Service]())
 	if err != nil {
-		return nil, fmt.Errorf("failed to add event handler: %w", err)
+		return nil, fmt.Errorf("failed to add indexer: %w", err)
 	}
 
 	services := fixturesSvcWithLabels(toLookup, underTest)