## What does it do ?

TODO:
- results from real clusters

This is a follow-up PR for https://github.com/kubernetes-sigs/external-dns/pull/5536

- Added indexers to speed up lookups
- Added transfomers to reduce memory footprint

Before

```
BenchmarkEndpointTargetsFromServicesMedium-16                       	   11194	    105419 ns/op
BenchmarkEndpointTargetsFromServicesMediumIterateOverGateways
BenchmarkEndpointTargetsFromServicesMediumIterateOverGateways-16    	     290	   4070484 ns/op
BenchmarkEndpointTargetsFromServicesHigh
BenchmarkEndpointTargetsFromServicesHigh-16                         	      88	  12669344 ns/op
BenchmarkEndpointTargetsFromServicesHighIterateOverGateways
BenchmarkEndpointTargetsFromServicesHighIterateOverGateways-16      	       1	13425263466 ns/op
```

With the change

```
BenchmarkEndpointTargetsFromServicesMedium
BenchmarkEndpointTargetsFromServicesMedium-14                       	  586600	      2069 ns/op
BenchmarkEndpointTargetsFromServicesMediumIterateOverGateways
BenchmarkEndpointTargetsFromServicesMediumIterateOverGateways-14    	    5715	    210000 ns/op
BenchmarkEndpointTargetsFromServicesHigh
BenchmarkEndpointTargetsFromServicesHigh-14                         	  540793	      1881 ns/op
BenchmarkEndpointTargetsFromServicesHighIterateOverGateways
BenchmarkEndpointTargetsFromServicesHighIterateOverGateways-14      	       6	 192442562 ns/op
```

This are syntetic benchmarks. Same time the difference is 5x. So it should provide a potential improvement in terms of cpu cycles and processing times

Follow-up:
- move optimisations for istio source; example from pod service https://github.com/kubernetes-sigs/external-dns/blob/468fb66758aa55e7e5ef48670e9db205a7e1e6f5/source/pod.go#L69.
- do not make API call for ingresses, instead use caching/indexing/transformations

## Motivation

PR https://github.com/kubernetes-sigs/external-dns/pull/5473
Issue #5458
Potential why performance is poor description https://github.com/kubernetes-sigs/external-dns/pull/5473#issuecomment-2911576025

## More

- [X] Yes, this PR title follows [Conventional Commits](https://www.conventionalcommits.org/en/v1.0.0/)
- [X] Yes, I added unit tests
- [ ] Yes, I updated end user documentation accordingly

<!--
    Please read https://github.com/kubernetes-sigs/external-dns#contributing before submitting
    your pull request. Please fill in each section below to help us better prioritize your pull request. Thanks!
-->