@ivankatliarchuk 2025-08-02
I'll block this. I think found a bug or issue that this approach may not work as expected

/hold