@ivankatliarchuk 2025-08-02
Great example of unit-tests all green, but reality is different. The case I found did not have any tests coverage.